records: move populate_earliest_date to record utils
