+++ tests/unit/records/test_records_receivers.py
@@ -34,7 +34,6 @@ from inspirehep.modules.records.receivers import (
     populate_affiliation_suggest,
     populate_authors_name_variations,
     populate_bookautocomplete,
-    populate_earliest_date,
     populate_experiment_suggest,
     populate_inspire_document_type,
     populate_recid_from_ref,
@@ -586,130 +585,6 @@ def test_assign_uuid_does_not_touch_existing_uuids(mock_uuid4):
 
     assert validate(result, subschema) is None
     assert expected == result
-
-
-def test_populate_earliest_date_from_preprint_date():
-    schema = load_schema('hep')
-    subschema = schema['properties']['preprint_date']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'preprint_date': '2014-05-29',
-    }
-    assert validate(record['preprint_date'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2014-05-29'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_from_thesis_info_date():
-    schema = load_schema('hep')
-    subschema = schema['properties']['thesis_info']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'thesis_info': {
-            'date': '2008',
-        },
-    }
-    assert validate(record['thesis_info'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2008'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_from_thesis_info_defense_date():
-    schema = load_schema('hep')
-    subschema = schema['properties']['thesis_info']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'thesis_info': {
-            'defense_date': '2012-06-01',
-        },
-    }
-    assert validate(record['thesis_info'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2012-06-01'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_from_publication_info_year():
-    schema = load_schema('hep')
-    subschema = schema['properties']['publication_info']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'publication_info': [
-            {'year': 2014},
-        ],
-    }
-    assert validate(record['publication_info'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2014'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_from_legacy_creation_date():
-    schema = load_schema('hep')
-    subschema = schema['properties']['legacy_creation_date']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'legacy_creation_date': '2015-11-04',
-    }
-    assert validate(record['legacy_creation_date'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2015-11-04'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_from_imprints_date():
-    schema = load_schema('hep')
-    subschema = schema['properties']['imprints']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'imprints': [
-            {'date': '2014-09-26'},
-        ],
-    }
-    assert validate(record['imprints'], subschema) is None
-
-    populate_earliest_date(record)
-
-    expected = '2014-09-26'
-    result = record['earliest_date']
-
-    assert expected == result
-
-
-def test_populate_earliest_date_does_nothing_if_record_is_not_literature():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_earliest_date(record)
-
-    assert 'earliest_date' not in record
 
 
 def test_populate_inspire_document_type_from_document_type():
